Move out the groups local servlets to a separate file.
